Fix workflow of non geostory resources
