+++ web/client/utils/MediaEditorUtils.js
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2019, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+/**
+ * Utils for media editor
+ */
+
+export const SourceTypes = {
+    GEOSTORY: 'geostory',
+    GEOSTORE: 'geostore'
+};
+export const defaultLayerMapPreview = {
+    type: 'osm',
+    title: 'Open Street Map',
+    name: 'mapnik',
+    source: 'osm',
+    group: 'background',
+    visibility: true,
+    id: 'mapnik__0',
+    loading: false,
+    loadingError: false
+};
